+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+* Fix issue where webhook data passed to `LocalPaymentReversed` webhook class is missing
+
 ## 3.5.0
 * Add `paymentReaderCardDetails` parameter to `Transaction.sale`
 * Add `skipAdvancedFraudChecking` to `Customer.create` and `Customer.update`